Refactored route creator - added few changes to log,cog,initial points

+++ src/main/java/com/wicio/shiplog/log/api/dto/CreateLogRequest.java
@@ -7,7 +7,7 @@ public record CreateLogRequest(
     Double YCoordinate,
     Double XCoordinate,
     Double speedOverGround,
-    Double courseOverGround,
+    Integer courseOverGround,
     Double windDirection,
     Double windSpeed,
     Boolean stationary) {
